FPM-32 Moved class Authentication from RegisterRequest to controller, implemented AuthenticationResponse methods
